Refactor types
