+++ packages/upswyng-types/src/TUser.ts
@@ -1,0 +1,8 @@
+export interface TUser {
+  _id: string; // database ObjectId converted to hex string
+  name?: string;
+  email: string;
+  providers: ("facebook" | "google" | "slack")[];
+  isAdmin: boolean;
+  isSuperAdmin: boolean;
+}
